test-requirements.txt
-r requirements.txt

coverage=4.0.3
django-debug-toolbar==1.4
